// src/types/profile.ts
export interface Education {
  schoolName: string;
  major: string;
  degree: string;
  admissionDate: string;
  graduationDate: string;
  educationStatus: string;
}

export interface Careers {
  companyName: string;
  department: string;
  position: string;
  startedAt: string;
  endedAt: string;
}